+++ R/chk-type.R
@@ -4,6 +4,7 @@
 #'
 #' @inheritParams chk_true
 #' @param class A string specifying the class.
+#' @param formals A count of the number of formal arguments.
 #' @return The \code{chk_} functions throw an informative error if the test fails.
 #' The \code{vld_} functions return a flag indicating whether the test was met.
 #' @name chk_type
@@ -184,19 +185,20 @@ vld_list <- function(x) is.list(x)
 #' # chk_function
 #' chk_function(mean)
 #' try(chk_function(1))
-chk_function <- function(x, x_name = NULL) {
-  if (vld_function(x)) {
+chk_function <- function(x, formals = NULL, x_name = NULL) {
+  if (vld_function(x, formals)) {
     return(invisible())
   }
   if (is.null(x_name)) x_name <- deparse_backtick_chk(substitute(x))
-  abort_chk(x_name, " must be a function")
+  if(!is.function(x)) abort_chk(x_name, " must be a function")
+  abort_chk(x_name, " must have ", formals, " formal arguments (not ", length(formals(x)), ")")
 }
 
 #' @describeIn chk_type Validate Function
 #'
 #' Validates is a function using:
 #'
-#' \code{\link{is.function}(x)}
+#' \code{\link{is.function}(x) && (is.null(formals) || length(formals(x)) == formals)}
 #'
 #' @export
 #'
@@ -207,7 +209,9 @@ chk_function <- function(x, x_name = NULL) {
 #' vld_function(function(x) x)
 #' vld_function(1)
 #' vld_function(list(1))
-vld_function <- function(x) is.function(x)
+vld_function <- function(x, formals = NULL) {
+  is.function(x) && (is.null(formals) || length(formals(x)) == formals)
+}
 
 #' @describeIn chk_type Check Vector
 #'